refactor: move and rename utils file
